feature: off api 차량정보 누적거리 수정, 운행일지 생성

+++ monicar-event-hub/src/main/java/org/eventhub/domain/VehicleInformation.java
@@ -18,7 +18,7 @@ public class VehicleInformation {
 	private Integer pv;
 	private Integer did;
 	private Integer drivingDays;
-	private Integer sum;
+	private Long sum;
 	private VehicleStatus status;
 	private LocalDate deliveryDate;
 	private LocalDateTime createdAt;
